requirements.txt
anyio==3.2.1
asyncclick==7.1.2.3
certifi==2021.5.30
chardet==4.0.0
docutils==0.17.1
idna==2.10
importlib-metadata==4.6.0
Kivy==2.0.0
Kivy-Garden==0.1.4
Pygments==2.9.0
python-kasa==0.4.0.dev3
requests==2.25.1
sniffio==1.2.0
urllib3==1.26.6
zipp==3.4.1
